feat: Add data collection mode and integrate CNN
This commit introduces a new "collect" mode for data collection, which saves sensor data to an HDF5 file. The "online" mode has been removed, and the "nav" and "gps" modes have been updated to use a CNN for angle prediction. The angle printing has also been standardized to use degrees.